Test pages

+++ main/bsp/temperature.h
@@ -2,9 +2,12 @@
 #define BSP_TEMPERATURE_H_INCLUDED
 
 
-void   bsp_temperature_init(void);
-double bsp_temperature_manage(void);
-double bsp_temperature_get(void);
+#include <stdint.h>
+
+
+void    bsp_temperature_init(void);
+uint8_t bsp_temperature_manage(void);
+double  bsp_temperature_get(void);
 
 
 #endif
